refactor: consolidate command architecture and update stale README
- Consolidate /gh_prs into /gh_pr list subcommand
- Consolidate /gh_issues into /gh_issue list subcommand
- Update README with accurate command syntax and examples
- Remove auto-unfurl section (feature not implemented)
- Maintain full backward compatibility for existing commands
- Merge tests into unified handler test files

Commands now follow consistent pattern:
- /gh_pr owner/repo #123 [--full] - show single PR
- /gh_pr list owner/repo [count] [filters] - list PRs
- /gh_issue owner/repo #123 [--full] - show single issue
- /gh_issue list owner/repo [count] [filters] - list issues

Breaking changes: None
- Old /gh_pr and /gh_issue commands work exactly as before
- New list functionality added via subcommand pattern
